--- lib/OrdinaryDiffEqDAE/src/OrdinaryDiffEqDAE.jl
@@ -1,20 +1,0 @@
-module OrdinaryDiffEqDAE
-
-import OrdinaryDiffEq: _unwrap_val, NLNewton, DAEAlgorithm,
-                       DEFAULT_PRECS, OrdinaryDiffEqConstantCache,
-                       OrdinaryDiffEqMutableCache, OrdinaryDiffEqConstantCache,
-                       @unpack, @cache, default_controller, stepsize_controller!, 
-                       step_accept_controller!,
-                       step_reject_controller!, post_newton_controller!
-using TruncatedStacktraces, MuladdMacro
-import FastBroadcast: @..
-
-include("algorithms.jl")
-include("alg_utils.jl")
-include("controllers.jl")
-include("dae_caches.jl")
-include("dae_perform_step.jl")
-
-export DABDF2, DImplicitEuler, DFBDF
-
-end # module OrdinaryDiffEqDAE
